refactor: stage 1
