+++ src/constants/interfaces.tsx
@@ -18,11 +18,18 @@ export interface AppContextComponentProps {
   children: ReactNode;
 }
 
+export interface ModalContextComponentProps {
+  children: ReactNode;
+}
+
 export interface AppContextProps {
   currentTopSection: string;
   onChange: Dispatch<SetStateAction<string>>;
-  onModalContentChange: Dispatch<SetStateAction<ReactNode>>;
   contactSectionRef: MutableRefObject<HTMLElement | null>;
+}
+
+export interface ModalContextProps {
+  onModalContentChange: Dispatch<SetStateAction<ReactNode>>;
   modalRef: MutableRefObject<HTMLDialogElement | null>;
   modalContent: ReactNode;
   onCloseModal: (e: Event | React.MouseEvent<HTMLElement>) => void;
